src/requirements.txt
requests
flask
pychart.js
python-dotenv
Flask-SQLAlchemy
sqlalchemy
pytz
flask-paginate
python-dateutil
psycopg2
musicbrainzngs
pytest
pytest-mock